Adds full creating of a team

+++ PEClient/Models/Student.cs
@@ -39,6 +39,6 @@ namespace PEClient.Models
     {
         public string Name;
         public string UserName;
-        public int id;
+        public decimal id;
     }
 }
